requirements.txt
faiss-cpu
sentence-transformers
langchain
langchain-community
pydantic
pyyaml
transformers
accelerate
torch
ollama
requests